# DROOLS-5421 DMN FEEL perf opt EvaluationContext.current()